+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
-## 0.37.0
-* `MapMemory` now implements `serde::Serialize` and `serde::Deserialize` when the `serde` feature is enabled
+## Unreleased
+
+* `MapMemory` now implements `serde::Serialize` and `serde::Deserialize` when the `serde` feature is
+  enabled.
 
 ## 0.36.0